transcript/requirements.txt
psycopg2-binary==2.9.3
requests==2.27.1
wget==3.2